FEAT: 리뷰 리스트 조회, 리뷰 자세히보기 기능 추가

+++ src/main/java/com/example/reviews3/dto/ReviewRequestDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Builder
 public class ReviewRequestDTO {
 	private String content;
-	private Long rating;
+	private Float rating;
 	private List<MultipartFile> image;
 
 	public Review toEntity(){
